milestone 3/pyproject.toml
[project]
name = "milestone-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=3.6.0",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-huggingface>=0.2.0",
    "langchain-openai>=0.3.16",
    "python-dotenv>=1.1.0",
    "rouge-score>=0.1.2",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.1",
    "uvicorn>=0.34.2",
]
